+++ solidblocks-rds-postgresql-hetzner/snippets/hetzner-postgres-rds-hetzner-s3-backup/instance/main.tf
@@ -13,7 +13,7 @@ resource "hcloud_ssh_key" "ssh_key" {
 }
 
 module "rds-postgresql" {
-  source = "https://github.com/pellepelster/solidblocks/releases/download/v0.4.8/blcks-terraform-rds-postgresql-hetzner-v0.4.8.zip"
+  source = "https://github.com/pellepelster/solidblocks/releases/download/v0.4.9/blcks-terraform-rds-postgresql-hetzner-v0.4.9.zip"
 
   name     = "rds-postgresql"
   location = var.hetzner_location