+++ solidblocks-rds-postgresql-hetzner/snippets/hetzner-postgres-rds-local-backup/instance/main.tf
@@ -17,7 +17,7 @@ data "hcloud_volume" "backup" {
 }
 
 module "rds-postgresql" {
-  source = "https://github.com/pellepelster/solidblocks/releases/download/v0.4.8/blcks-terraform-rds-postgresql-hetzner-v0.4.8.zip"
+  source = "https://github.com/pellepelster/solidblocks/releases/download/v0.4.9/blcks-terraform-rds-postgresql-hetzner-v0.4.9.zip"
 
   name     = "rds-postgresql"
   location = var.hetzner_location